+++ src/routes/guide/Guide.svelte
@@ -17,6 +17,7 @@
     } from '@concepts/ConceptParams';
     import { locales } from '@db/Database';
     import Project from '@db/projects/Project';
+    import { toLocale } from '@locale/LocaleText';
     import Source from '@nodes/Source';
     import { onMount } from 'svelte';
     import { writable } from 'svelte/store';
@@ -24,9 +25,7 @@
     function getLocaleInURL() {
         return (
             $page.url.searchParams.get('locale') ??
-            `${$locales.getLocales()[0].language}-${
-                $locales.getLocales()[0].region
-            }`
+            toLocale($locales.getLocales()[0])
         );
     }